refactor(container): refacto pnr code
refactor to improve code quality and performance

ref: MANAGER-14777
